requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pyjokes
newsapi-python